Moved some files and namespaces around
