+++ Thinktecture.Relay.Server/Controller/ClientController.cs
@@ -9,7 +9,7 @@ using Thinktecture.Relay.Server.Diagnostics;
 using Thinktecture.Relay.Server.Dto;
 using Thinktecture.Relay.Server.Helper;
 using Thinktecture.Relay.Server.Http;
-using Thinktecture.Relay.Server.Http.ActionFilters;
+using Thinktecture.Relay.Server.Http.Filters;
 using Thinktecture.Relay.Server.OnPremise;
 using Thinktecture.Relay.Server.Interceptor;
 using Thinktecture.Relay.Server.Repository;
