+++ Thinktecture.Relay.Server/Controller/ManagementWeb/DashboardController.cs
@@ -1,6 +1,6 @@
 using System.Web.Http;
 using Thinktecture.Relay.Server.Dto;
-using Thinktecture.Relay.Server.Http.ActionFilters;
+using Thinktecture.Relay.Server.Http.Filters;
 using Thinktecture.Relay.Server.Repository;
 
 namespace Thinktecture.Relay.Server.Controller.ManagementWeb
