+++ Thinktecture.Relay.Server/Controller/ManagementWeb/LogController.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Web.Http;
 using Thinktecture.Relay.Server.Dto;
-using Thinktecture.Relay.Server.Http.ActionFilters;
+using Thinktecture.Relay.Server.Http.Filters;
 using Thinktecture.Relay.Server.Repository;
 
 namespace Thinktecture.Relay.Server.Controller.ManagementWeb
