+++ Thinktecture.Relay.Server/Controller/ManagementWeb/SetupController.cs
@@ -1,7 +1,7 @@
 using System.Net;
 using System.Web.Http;
 using System.Web.Http.Results;
-using Thinktecture.Relay.Server.Http.ActionFilters;
+using Thinktecture.Relay.Server.Http.Filters;
 using Thinktecture.Relay.Server.Repository;
 
 namespace Thinktecture.Relay.Server.Controller.ManagementWeb
