+++ Thinktecture.Relay.Server/Controller/ManagementWeb/UserController.cs
@@ -4,7 +4,7 @@ using System.Net;
 using System.Web.Http;
 using System.Web.Http.Results;
 using Thinktecture.Relay.Server.Dto;
-using Thinktecture.Relay.Server.Http.ActionFilters;
+using Thinktecture.Relay.Server.Http.Filters;
 using Thinktecture.Relay.Server.Repository;
 
 namespace Thinktecture.Relay.Server.Controller.ManagementWeb
