+++ Thinktecture.Relay.Server/Controller/RequestController.cs
@@ -2,7 +2,7 @@ using System.Net.Http;
 using System.Web.Http;
 using System.Web.Http.Results;
 using Serilog;
-using Thinktecture.Relay.Server.Http.ActionFilters;
+using Thinktecture.Relay.Server.Http.Filters;
 using Thinktecture.Relay.Server.SignalR;
 
 namespace Thinktecture.Relay.Server.Controller
