+++ Thinktecture.Relay.Server/Thinktecture.Relay.Server.csproj
@@ -203,7 +203,7 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Http\ActionFilters\NoCacheAttribute.cs" />
+    <Compile Include="Http\Filters\NoCacheAttribute.cs" />
     <Compile Include="DependencyInjection\RelayServerModule.cs" />
     <Compile Include="Dto\LinkDetails.cs" />
     <Compile Include="IStartup.cs" />
@@ -231,9 +231,9 @@
     <Compile Include="Config\IPersistedSettings.cs" />
     <Compile Include="Config\LocalAppDataPersistedSettings.cs" />
     <Compile Include="Config\ModuleBinding.cs" />
-    <Compile Include="Http\ActionFilters\CheckModuleBindingConfigurationAttribute.cs" />
-    <Compile Include="Http\ActionFilters\ManagementWebModuleBindingFilter.cs" />
-    <Compile Include="Http\ActionFilters\OnPremiseConnectionModuleBindingFilter.cs" />
+    <Compile Include="Http\Filters\CheckModuleBindingConfigurationAttribute.cs" />
+    <Compile Include="Http\Filters\ManagementWebModuleBindingFilter.cs" />
+    <Compile Include="Http\Filters\OnPremiseConnectionModuleBindingFilter.cs" />
     <Compile Include="Http\HttpRequestMessageExtensions.cs" />
     <Compile Include="Interceptor\IInterceptorLoader.cs" />
     <Compile Include="Interceptor\IInterceptorManager.cs" />
@@ -243,7 +243,7 @@
       <DependentUpon>201707240650482_AddActiveConnectionsTable.cs</DependentUpon>
     </Compile>
     <Compile Include="Repository\DbModels\DbActiveConnection.cs" />
-    <Compile Include="Http\ActionFilters\RelayModuleBindingFilter.cs" />
+    <Compile Include="Http\Filters\RelayModuleBindingFilter.cs" />
     <Compile Include="Interceptor\InterceptorLoader.cs" />
     <Compile Include="Interceptor\InterceptorManager.cs" />
     <Compile Include="Security\CustomJwtFormat.cs" />
